python/requirements.txt
python-dotenv
fastapi[standard-no-fastapi-cloud-cli]
pyserial==3.5
websockets==12.0
